% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_margins_cdl.R
\name{add_margins_cdl}
\alias{add_margins_cdl}
\title{Add margins to cropland data layer raster}
\usage{
add_margins_cdl(margin.lc = 171, margin.rast, cdl.rast)
}
\arguments{
\item{margin.lc}{margin habitat code (numeric value). Default is 171}

\item{margin.rast}{raster representing margin cells as non-NA values and all other cells NA}

\item{cdl.rast}{raster representing CDL land cover. Can use the \code{cdl.clean} layer from \code{make_fields} output}
}
\value{
raster of CDL land cover with field margins indicated by \code{margin.rast} with the \code{margin.lc} value
}
\description{
Add margin raster (output from \code{make_fieldmargins}) to a CDL layer
}
