% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fields.R
\name{make_fields}
\alias{make_fields}
\title{Make fields raster from crop sequence boundaries polygons}
\usage{
make_fields(fields.poly, field.col, cdl.rast, field.rcl, section.size)
}
\arguments{
\item{fields.poly}{Crop sequence boundaries polygons, in sf format}

\item{field.col}{Crop ID column name in csb polygon vector object}

\item{cdl.rast}{CDL raster on which to superimpose the \code{fields.poly} fields}

\item{field.rcl}{Reclass table to identify field cells}

\item{section.size}{length of field border to define as a section}
}
\value{
3-layer terra raster of field id, crop fields, and cdl with 'cleaned'
crop fields
}
\description{
Makes a layer with focal fields identified by the crop sequence boundary
layer. Outputs a 2-layer raster of the field cells only, with cell values
showing 1) field id, 2) crop fields, and 3) a 'cleaned' version of the cdl
where crop fields contain only a single crop type.  The function assigns a
sequential field id based on CSB polygon row number.
}
