% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fieldsections.R
\name{make_fieldsections}
\alias{make_fieldsections}
\title{Make field sections}
\usage{
make_fieldsections(field.poly, section.size, simplify.param = 100)
}
\arguments{
\item{field.poly}{a single field (sf object)}

\item{section.size}{field margin section size}

\item{simplify.param}{parameter for simplifying field polygon in intermediate
step. Default is 100.}
}
\value{
an sf object
}
\description{
Make field sections
}
