% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kc_curves.R
\name{make_kc_curves}
\alias{make_kc_curves}
\title{Make Kc curves}
\usage{
make_kc_curves(
  ini,
  mid,
  end,
  Lini,
  Ldev,
  Lmid,
  Llate,
  dayPlant,
  dayHarvest,
  doublecrop = FALSE,
  winter.dayPlant,
  winter.ini,
  winter.mid,
  winter.end,
  winter.Lini,
  winter.Ldev,
  winter.Lmid,
  winter.Llate,
  monthly = TRUE,
  Nistor = FALSE,
  cold
)
}
\arguments{
\item{ini}{Kc_ini crop coefficient}

\item{mid}{Kc_mid crop coefficient}

\item{end}{Kc_end crop coefficient}

\item{Lini}{L_ini length of Kc_ini period}

\item{Ldev}{L_dev length of development period between Kc_ini and Kc_mid
periods}

\item{Lmid}{L_mid length of Kc_mid period}

\item{Llate}{L_late length of period transitioning from end of Kc_mid to
Kc_end}

\item{dayPlant}{Julian day of spring crop}

\item{dayHarvest}{(Optional) date of harvest. Only needed if Lmid and Llate
are not provided}

\item{doublecrop}{T/F whether to calculate curve for a winter crop
(double-cropping), default is FALSE}

\item{winter.dayPlant}{(Optional) Julian day of winter planting, only for
double crops}

\item{winter.ini}{(Optional) Kc_ini for winter crop}

\item{winter.mid}{(Optional) Kc_mid for winter crop}

\item{winter.end}{(Optional) Kc_end for winter crop}

\item{winter.Lini}{(Optional) L_ini length of Kc_ini of winter crop}

\item{winter.Ldev}{(Optional) L_dev length of development between Kc_ini and
Kc_mid of winter crop}

\item{winter.Lmid}{(Optional) L_mid length of Kc_mid period of winter crop}

\item{winter.Llate}{(Optional) L_late length of period from end of Kc_mid to
Kc_end of winter crop}

\item{monthly}{T/F whether to return daily (F) or monthly (T) Kc values.}

\item{Nistor}{T/F whether to use the method in Nistor et al. 2018}

\item{cold}{cold period parameter from Nistor et al. 2018}
}
\value{
A dataframe of daily or monthly Kc values based on inputs
}
\description{
Follows the FAO method of defining Kc curves based on crop development
stages. Kc values are calculated for every day of the year based on the
user-provided Kc values, development dates, and stage lengths from Allen et
al. (1998). If \code{doublecrop} is indicated, then these parameter values for the
second crop must also be provided. If \code{Nistor} is TRUE, then season-based Kc
values from Nistor et al. (2018) need to be provided.
}
\examples{
make_kc_curves(ini=0.7,mid=1.2,end=0.48,Lini=30,Ldev=40,Lmid=50,Llate=50,dayPlant=105,doublecrop=TRUE,winter.dayPlant=288,winter.ini=0.4,winter.mid=1.15,winter.end=0.33,winter.Lini=30,winter.Ldev=140,winter.Lmid=40,winter.Llate=30)

}
