% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cropbdry.R
\name{query_cropbdry}
\alias{query_cropbdry}
\title{Return crop sequence boundary polygons falling within a polygon}
\usage{
query_cropbdry(polysf, cropbdry.path)
}
\arguments{
\item{polysf}{sf polyon}

\item{cropbdry.path}{Character string to cropbdry geodatabase}
}
\value{
sf polygon object of crop sequence boundary polygons falling within
\code{polysf}
}
\description{
Return crop sequence boundary polygons falling within a polygon
}
