% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fieldmargins.R
\name{make_fieldmargins}
\alias{make_fieldmargins}
\title{Make field margin layers}
\usage{
make_fieldmargins(field.id.rast, field.crop.rast, show.progress = TRUE)
}
\arguments{
\item{field.id.rast}{field id raster, \code{id} layer returned from \code{make_fields}}

\item{field.crop.rast}{field crop raster, \code{crop} layer returned from
\code{make_fields}}

\item{show.progress}{T/F show progress}
}
\value{
A list with a raster stack of 1) margin cell id, 2) margin cell cost,
and 3) margin field id
}
\description{
Takes field polygons and overlays them on land cover layer (cdl) and
determines field margins, including inner and outer margin costs
}
